Create functions.cpp
This is pretty self explanatory if you know what is called first since int main(). It's checkSentence().

checkSentence has a double for loop that loops through sentence trying to match our "No No" word (word) with any position in sentence.

If the number of (ifMatches) ever matches the size of word then bleep is called. To take that position[i] and change every position to a " * ".

once everything is finished since sentence was globally referenced there is no need to return any values or print from the functions. You can do that from the main().